backend/account-service/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
httpx